+++ R/Plotting_Functions.R
@@ -13,7 +13,7 @@
 #'
 #' Points can also be sized according to their associated clustering uncertainty with the option \code{"uncertainty"}. In doing so, the transparency of the points will also be proportional to their clustering uncertainty, provided the device supports transparency. See also \code{\link{MoE_Uncertainty}} for an alternative means of visualising observation-specific cluster uncertainties (especially for univariate data). See \code{scatter.pars} below, and note that models fitted via the \code{"CEM"} algorithm will have no associated clustering uncertainty.
 #'
-#' Alternatively, the bivariate \code{"density"} contours can be displayed (see \code{density.pars}), provided there is at least one Gaussian component in the model. Caution is advised when producing density plots for models with covariates in the expert network; the required number of evaluations of the (multivariate) Gaussian density for each panel (\code{res$G * prod(density.pars$grid.size)}) increases by a factor of \code{res$n}, thus plotting may be slow (particularly for large data sets). See \code{density.pars} below.
+#' Alternatively, the bivariate \emph{parametric} \code{"density"} contours can be displayed (see \code{density.pars}), provided there is at least one Gaussian component in the model. Caution is advised when producing density plots for models with covariates in the expert network; the required number of evaluations of the (multivariate) Gaussian density for each panel (\code{res$G * prod(density.pars$grid.size)}) increases by a factor of \code{res$n}, thus plotting may be slow (particularly for large data sets). See \code{density.pars} below.
 #' @param scatter.type A vector of length 2 (or 1) giving the plot type for the upper and lower triangular portions of the plot, respectively, pertaining to the associated covariates. Defaults to \code{"lm"} for covariate vs. response panels and \code{"points"} otherwise. Only relevant for models with continuous covariates in the gating &/or expert network. \code{"ci"} and \code{"lm"} type plots are only produced for plots pairing covariates with response, and never response vs. response or covariate vs. covariate. Note that lines &/or confidence intervals will only be drawn for continuous covariates included in the expert network; to include covariates included only in the gating network also, the options \code{"lm2"} or \code{"ci2"} can be used but this is not generally advisable. See \code{scatter.pars} below.
 #' @param conditional A vector of length 2 (or 1) giving the plot type for the upper and lower triangular portions of the plot, respectively, for plots involving a mix of categorical and continuous variables. Defaults to \code{"stripplot"} in the upper triangle and \code{"boxplot"} in the lower triangle (see \code{\link[lattice]{panel.stripplot}} and \code{\link[lattice]{panel.bwplot}}). \code{"violin"} and \code{"barcode"} plots can also be produced. Only relevant for models with categorical covariates in the gating &/or expert network, unless \code{show.MAP} is \code{TRUE}. Comparisons of two categorical variables (which can only ever be covariates or the MAP classification) are always displayed via mosaic plots (see \code{\link[vcd]{strucplot}}). 
 #' 
@@ -45,12 +45,12 @@
 #' where \code{scat.pch}, \code{scat.col}, and \code{scat.size} give the plotting symbols, colours, and sizes of the points in scatterplot panels, respectively. Note that \code{eci.col} gives both a) the colour of the fitted lines &/or confidence intervals for expert-related panels when \code{scatter.type} is one of \code{"ci"} or \code{"lm"} and b) the colour of the ellipses (if any) when \code{addEllipses} is one of \code{"outer"}, \code{"inner"}, or \code{"both"} and the response data is multivariate. Note that \code{eci.col} will inherit a suitable default from \code{scat.col} instead if the latter is supplied but the former is not. 
 #' 
 #' Note also that \code{scat.size} will be modified on an observation-by-observation level when \code{response.type} is \code{"uncertainty"}. Furthermore, note that the behaviour for plotting symbols when \code{response.type="uncertainty"} changes compared to \code{response.type="points"} depending on the value of the \code{uncert.cov} argument above. \code{uncert.pch} gives the plotting symbol used for all scatterplot (and strip plot) panels when \code{response.type="uncertainty"} and \code{uncert.cov} is \code{TRUE}. However, when \code{uncert.cov} is \code{FALSE}, \code{scat.pch} is invoked for scatterplots involving covariates and \code{uncert.pch} is used for panels involving only response variables. Finally, \code{noise.size} can be used to modify \code{scat.size} for observations assigned to the noise component (if any), but only when \code{response.type="points"}.
-#' @param density.pars A list supplying select parameters for visualising the bivariate density contours, only when \code{response.type} is \code{"density"}.
+#' @param density.pars A list supplying select parameters for visualising the bivariate \emph{parametric} density contours, only when \code{response.type} is \code{"density"}.
 #'
 #' \code{NULL} is equivalent to:
-#' \preformatted{list(grid.size=c(100, 100), dcol="grey50",
-#'      nlevels=11, show.labels=TRUE, label.style="mixed"),}
-#' where \code{grid.size} is a vector of length two giving the number of points in the x & y direction of the grid over which the density is evaluated, respectively (though \code{density.pars$grid.size} can also be supplied as a scalar, which will be automatically recycled to a vector of length 2), and \code{dcol} is either a single colour or a vector of length \code{nlevels} colours (although note that \code{dcol}, when \emph{not} specified, will be adjusted for transparency). Finally, \code{label.style} can take the values \code{"mixed"}, \code{"flat"}, or \code{"align"}. Note that \code{density.pars$grid.size[1]} is also relevant when \code{diag.pars$show.dens=TRUE} (see below).
+#' \preformatted{list(grid.size=c(100, 100), dcol="grey50", dens.points=FALSE,
+#'      nlevels=11, show.labels=!dens.points, label.style="mixed"),}
+#' where \code{grid.size} is a vector of length two giving the number of points in the x & y directions of the grid over which the density is evaluated, respectively (though \code{density.pars$grid.size} can also be supplied as a scalar, which will be automatically recycled to a vector of length 2), and \code{dcol} is either a single colour or a vector of length \code{nlevels} colours. \code{dens.points} governs whether points should be overlaid when \code{response.type="density"} (in other words, \code{dens.points=TRUE} is akin to specifying \code{response.type="points"} and \code{response.type="density"} simultaneously) and \code{show.labels} governs whether the density contours should be labelled. Note that contours are not labelled when \code{dens.points=TRUE} by default. Finally, \code{label.style} can take the values \code{"mixed"}, \code{"flat"}, or \code{"align"}.
 #' @param stripplot.pars A list supplying select parameters for continuous vs. categorical panels when one or both of the entries of \code{conditional} is \code{"stripplot"}.
 #'
 #' \code{NULL} is equivalent to:
@@ -83,11 +83,11 @@
 #' @param diag.pars A list supplying select parameters for panels along the diagonal.
 #'
 #' \code{NULL} is equivalent to:
-#' \preformatted{list(diag.fontsize=9, show.hist=TRUE, show.dens=FALSE,
-#'      diagonal=TRUE, hist.color=hist.color, show.counts=TRUE),}
+#' \preformatted{list(diag.fontsize=9, diagonal=TRUE, hist.color=hist.color, 
+#'                    show.hist=TRUE, show.counts=TRUE, show.dens=FALSE, dens.grid=100),}
 #' where \code{hist.color} is a vector of length 4, giving the colours for the response variables, gating covariates, expert covariates, and covariates entering both networks, respectively. By default, diagonal panels for response variables are \code{ifelse(diag.pars$show.dens, "white", "black")} and covariates of any kind are \code{"dimgrey"}. \code{hist.color} also governs the outer colour for mosaic panels and the fill colour for boxplot, violin, and barcode panels (except for those involving the MAP classification). However, in the case of response vs. (categorical) covariates boxplots and violin plots, the fill colour is always \code{"white"}. The MAP classification is always coloured by cluster membership, by default. The argument \code{show.counts} is only relevant for categorical variables.
 #'
-#' The argument \code{show.dens} toggles whether parametric density estimates are drawn over the diagonal panels for each response variable. When \code{show.dens=TRUE}, the component densities are shown via thin lines, with colours given by \code{scatter.pars$scat.col}, while a thick \code{"black"} line is used for the overall mixture density. This argument can be used with or without \code{show.hist} also being \code{TRUE}. Finally, the grid size when \code{show.dens=TRUE} is given by \code{max(res$n, density.pars$grid.size[1])}.
+#' The argument \code{show.dens} toggles whether \emph{parametric} density estimates are drawn over the diagonal panels for each response variable. When \code{show.dens=TRUE}, the component densities are shown via thin lines, with colours given by \code{scatter.pars$scat.col}, while a thick \code{"black"} line is used for the overall mixture density. This argument can be used with or without \code{show.hist} also being \code{TRUE}. Finally, the grid size when \code{show.dens=TRUE} is given by \code{diag.grid=100} by default.
 #' 
 #' When \code{diagonal=TRUE} (the default), the diagonal from the top left to the bottom right is used for displaying the marginal distributions of variables (via histograms, with or without overlaid density estimates, or barplots, as appropriate). Specifying \code{diagonal=FALSE} will place the diagonal running from the top right down to the bottom left.
 #' @param ... Catches unused arguments. Alternatively, named arguments can be passed directly here to any/all of \code{scatter.pars}, \code{stripplot.pars}, \code{boxplot.pars}, \code{barcode.pars}, \code{mosaic.pars}, \code{axis.pars}, and \code{diag.pars}.
@@ -159,15 +159,17 @@
 #' # Use different colours for histograms of covariates in the gating/expert/both networks.
 #' # Also use different colours for response vs. covariate & covariate vs. response panels.
 #' 
-#' MoE_gpairs(res, response.type="density", show.labels=FALSE,
+#' MoE_gpairs(res, response.type="density", show.labels=FALSE, dens.points=TRUE,
 #'            hist.color=c("black", "cyan", "hotpink", "chartreuse"),
 #'            bg.col=c("whitesmoke", "white", "mintcream", "mintcream", "floralwhite"))
 #'            
-#' # Examine the effect of the expert.covar argument in conjunction with show.dens & show.hist
-#' MoE_gpairs(res, cov.ind=0, expert.covar=TRUE, 
-#'            show.dens=TRUE, show.hist=FALSE, grid.size=1000)
+#' # Examine effect of expert.covar & diag.grid in conjunction with show.dens & show.hist
 #' MoE_gpairs(res, cov.ind=0, expert.covar=FALSE, 
-#'            show.dens=TRUE, show.hist=TRUE, grid.size=1000)
+#'            show.dens=TRUE, show.hist=FALSE, diag.grid=20)
+#' MoE_gpairs(res, cov.ind=0, expert.covar=TRUE, 
+#'            show.dens=TRUE, show.hist=TRUE, diag.grid=200)
+#' MoE_gpairs(res, cov.ind=0, expert.covar=TRUE, 
+#'            show.dens=FALSE, show.hist=FALSE, diagonal=FALSE)          
 #'            
 #' # Produce a generalised pairs plot for a model with a noise component.
 #' # Reorder the covariates and omit the variables "Hc" and "Hg".
@@ -423,13 +425,17 @@ MoE_gpairs.MoEClust <- function(res, response.type = c("points", "uncertainty", 
             !is.numeric(density.pars$nlevels)    ||
             density.pars$nlevels            <= 1)     stop("Invalid 'density.pars$nlevels'", call.=FALSE)
   if(is.null(density.pars$dcol))     {
-    density.pars$dcol    <- if(grDevices::dev.capabilities()$semiTransparency) unname(mapply(grDevices::adjustcolor, col="grey50", alpha.f=seq(0.5, 1, length=density.pars$nlevels))) else "grey50"
+    density.pars$dcol    <- "grey50"
   } else if(!is.element(length(density.pars$dcol),
             c(1, density.pars$nlevels))     ||
             !is.character(density.pars$dcol))         stop("Invalid 'density.pars$dcol", call.=FALSE)
   density.pars$dcol      <- if(length(density.pars$dcol) == 1) rep(density.pars$dcol, density.pars$nlevels) else density.pars$dcol
+  if(is.null(density.pars$dens.points)) {
+    density.pars$dens.points <- FALSE
+  } else if(length(density.pars$dens.points) > 1 ||
+            !is.logical(density.pars$dens.points))    stop("Invalid 'density.pars$dens.points", call.=FALSE)
   if(is.null(density.pars$show.labels))  {
-    density.pars$show.labels <- TRUE
+    density.pars$show.labels <- !density.pars$dens.points
   } else if(length(density.pars$show.labels) > 1 ||
             !is.logical(density.pars$show.labels))    stop("Invalid 'density.pars$show.labels", call.=FALSE)
   if(is.null(density.pars$label.style))  {
@@ -462,7 +468,9 @@ MoE_gpairs.MoEClust <- function(res, response.type = c("points", "uncertainty", 
   if(is.null(diag.pars$show.dens))   {
     show.D <- FALSE
   } else {
-    show.D <- diag.pars$show.dens
+    if(length(diag.pars$show.dens)   > 1 ||
+       !is.logical(diag.pars$show.dens))              stop("'diag.pars$show.dens' must be a single logical indicator", call.=FALSE)
+    show.D <- isTRUE(diag.pars$show.dens)
   }
   if(response.type == "density"   || show.D) {
     res$parameters$fits <- if(attr(res, "Expert"))  array(unlist(lapply(res$expert, "[[", "fitted.values")), dim=c(res$n, res$d, G))
@@ -470,7 +478,11 @@ MoE_gpairs.MoEClust <- function(res, response.type = c("points", "uncertainty", 
                               !attr(res, "Gating")) log(res$parameters$pro)
   }
   if(isTRUE(show.D)) {
-    diag.pars$grid.size <- density.pars$grid.size[1L]
+    if(is.null(diag.pars$diag.grid)) {
+      diag.pars$diag.grid   <- 100
+    }
+    if(length(diag.pars$diag.grid)  != 1 || !is.numeric(diag.pars$diag.grid) ||
+       diag.pars$diag.grid   < 10)                    stop("Invalid 'diag.pars$diag.grid'", call.=FALSE)
   }
   res$parameters$varianceX        <- if(isTRUE(expert.covar)  &&
                                         attr(res, "Expert")   &&
@@ -658,10 +670,10 @@ MoE_gpairs.MoEClust <- function(res, response.type = c("points", "uncertainty", 
           }
         }
         if(!any(is.factor(x[,i]), is.factor(x[,j])))  {
-          if(response.type == "density" && all(j <= dcol, i <= dcol)) {
-            .density_panel(cbind(x[,j], x[,i]), dimens=c(subset$data.ind[j - subset$show.map], subset$data.ind[i - subset$show.map]), res, density.pars, axis.pars, xpos, ypos, buffer, outer.rot, bg, border)
+          scatter.pars$pch  <- if(response.type == "uncertainty" && (all(j <= dcol, i <= dcol) || isTRUE(uncert.cov))) scatter.pars$uncert.pch else scatter.pars$scat.pch
+          if(response.type  == "density" && all(j <= dcol, i <= dcol)) {
+            .density_panel(cbind(x[,j], x[,i]), dimens=c(subset$data.ind[j - subset$show.map], subset$data.ind[i - subset$show.map]), res, density.pars, axis.pars, xpos, ypos, buffer, outer.rot, bg, border, scatter.pars)
           } else {
-            scatter.pars$pch   <- if(response.type == "uncertainty" && (all(j <= dcol, i <= dcol) || isTRUE(uncert.cov))) scatter.pars$uncert.pch else scatter.pars$scat.pch
             .scatter_panel(x=x[,j], y=x[,i], type=ifelse(j > dcol   && i <= dcol, ifelse(upr.gate || (j %in% c(expx, both)), upr.exp, "points"), ifelse(j <= dcol && i <= dcol, ifelse(drawEllipses, "ellipses", "points"), ifelse(j <= dcol && (low.gate || (i %in% c(expx, both))), low.exp, "points"))),
                            scatter.pars=scatter.pars, axis.pars=axis.pars, xpos=xpos, ypos=ypos, buffer=buffer, z=z, G=G, res=res, dimens=c(subset$data.ind[j - subset$show.map], subset$data.ind[i - subset$show.map]), outer.rot=outer.rot, bg=bg, mvn.type=addEllipses, border=border,
                            uncertainty=if(response.type == "uncertainty" && (uncert.cov  || (i <= dcol && j <= dcol))) uncertainty else NA, mvn.col=if(colEllipses) scatter.pars$eci.col)
@@ -1297,7 +1309,7 @@ plot.MoEClust <- function(x, what=c("gpairs", "gating", "criterion", "loglik", "
         ry     <- range(uy)
         depth  <- pmin(pmin(val$x - rx[1L], rx[2L] - val$x)/diff(rx), pmin(val$y - ry[1L], ry[2L] - val$y)/diff(ry))
         if(labels$style  == "align" | 
-           depth[txtloc  <- which.min(abs(slopes))] < 0.5) {
+           depth[txtloc  <- which.min(abs(slopes))] < 0.05) {
           txtloc         <- min(which.max(depth), length(slopes))
           rotang         <- atan(asp * slopes[txtloc] * diff(rx)/diff(ry)) * 180/base::pi  
         } else rotang    <- 0
@@ -1310,7 +1322,7 @@ plot.MoEClust <- function(x, what=c("gpairs", "gating", "criterion", "loglik", "
 
 #' @importFrom mclust "logsumexp"
 #' @importFrom matrixStats "colMeans2"
-.density_panel <- function(dat, dimens, res, density.pars, axis.pars, xpos, ypos, buffer, outer.rot, bg, border) {
+.density_panel <- function(dat, dimens, res, density.pars, axis.pars, xpos, ypos, buffer, outer.rot, bg, border, scatter.pars) {
   pars         <- res$parameters
   modelName    <- res$modelName
   G            <- res$G
@@ -1359,6 +1371,7 @@ plot.MoEClust <- function(x, what=c("gpairs", "gating", "criterion", "loglik", "
   grid::pushViewport(grid::viewport(xscale=xlim, yscale=ylim, clip=TRUE))
   grid::grid.rect(gp=grid::gpar(fill=bg, col=border))
   .contour_panel(x, y, zz, density.pars)
+  if(isTRUE(density.pars$dens.points)) grid::grid.points(dat[,1L], dat[,2L], pch=scatter.pars$pch, size=scatter.pars$size, gp=grid::gpar(col=scatter.pars$col, cex=1))
   grid::popViewport(1)
 }
 
@@ -1415,15 +1428,10 @@ plot.MoEClust <- function(x, what=c("gpairs", "gating", "criterion", "loglik", "
         sigma      <- c(sigma, list(sigmasq=pars$variance$sigma[i,i,]))
         sigma$modelName  <- "V"
       }
-      if((lx       <- diag.pars$grid.size) > n) {
-        rx         <- range(x, na.rm=TRUE)
-        xlim       <- rx + c(-buffer * (diff(rx)), buffer * (diff(rx)))
-        xd         <- .grid_1(n=lx, range=xlim, edge=TRUE)
-        xn         <- length(xd)  
-      } else {
-        xd         <- x
-        xn         <- n
-      }
+      lx           <- diag.pars$diag.grid
+      rx           <- range(x, na.rm=TRUE)
+      xlim         <- rx + c(-buffer * (diff(rx)), buffer * (diff(rx)))
+      xd           <- .grid_1(n=lx, range=xlim, edge=TRUE)
       Vinv         <- pars$Vinv
       noise        <- !is.null(Vinv)
       GN           <- G + noise
@@ -1441,9 +1449,9 @@ plot.MoEClust <- function(x, what=c("gpairs", "gating", "criterion", "loglik", "
       }   else den <- MoE_dens(data=xd,   mus=mu, sigs=sigma)
       if(isFALSE(gate))   {
         den        <- if(isTRUE(noise))    cbind(den, log(Vinv)) else den
-        den        <- den + .mat_byrow(pars$lpro, nrow=xn, ncol=GN)
+        den        <- den + .mat_byrow(pars$lpro, nrow=lx, ncol=GN)
       }
-      zz           <- matrix(exp(logsumexp(den)), nrow=xn, ncol=1L)
+      zz           <- matrix(exp(logsumexp(den)), nrow=lx, ncol=1L)
       den          <- exp(den)
       oo           <- order(xd)
       if(!diag.pars$show.hist) {